requirements.txt
requests
pymongo
django-crontab
subprocess
